api 연동 함수 및 type 컨트롤러별 분리, 유니온 타입 지정
