+++ packages/api/src/app/store/type.ts
@@ -1,0 +1,29 @@
+import { z } from 'zod';
+
+import { PageInfoSchema } from '../../shared';
+
+export const Store = z.object({
+  id: z.number(),
+  name: z.string(),
+  city: z.string(),
+  address: z.string(),
+  category: z.string(),
+  description: z.string(),
+  priceCategory: z.string(),
+  eventDescription: z.string(),
+  contactNumber: z.string(),
+  direction: z.string(),
+  information: z.string(),
+  isParkingAvailable: z.boolean(),
+  isNewOpen: z.boolean(),
+  isTakeOut: z.boolean(),
+});
+
+export type Store = z.infer<typeof Store>;
+
+export const GetStoresResponse = z.object({
+  list: z.array(Store),
+  pageInfo: PageInfoSchema,
+});
+
+export type GetStoresResponse = z.infer<typeof GetStoresResponse>;
